Added all of the files for Foundation
Cannot get the css file to link to the ftl  file. Need to find a
workaround.

Added all of the needed files.
Edited the home page some to figure out how it works.
The ftl file "user-home" is the page that gets loaded when
"/cs480/alextest" is loaded.
Possible icon for Shelter-cpp.